Major feature update
-separate accounts
-question scores are calculated differently
-quiz settings page with 3 modes
-can change category metadata
-list view updated

+++ app/src/main/java/m1/nayak/m1/ResultsFragment.java
@@ -78,7 +78,7 @@ public class ResultsFragment extends Fragment {
     public void getResults() {
         HashMap<String, Integer> correct = new HashMap<String, Integer>();
         HashMap<String, Integer> total = new HashMap<String, Integer>();
-        HashMap<String, Integer> score = new HashMap<String, Integer>();
+        HashMap<String, Double> score = new HashMap<String, Double>();
 
         for (int i = 0; i < Control.questions.size(); i++) {
             Log.d("ASH", "-----");
