+++ app/src/main/java/m1/nayak/m1/custom/ResultsAdapter.java
@@ -74,7 +74,7 @@ public class ResultsAdapter extends BaseAdapter {
     public void calculateResults() {
         HashMap<String, Integer> correct = new HashMap<String, Integer>();
         HashMap<String, Integer> total = new HashMap<String, Integer>();
-        HashMap<String, Integer> score = new HashMap<String, Integer>();
+        HashMap<String, Double> score = new HashMap<String, Double>();
         for (int i = 0; i < Control.questions.size(); i++) {
             Log.d("ASH", "-----");
             Log.d("ASH", "Subject = " + Control.questions.get(i).subject);
