+++ app/src/main/java/m1/nayak/m1/objects/FlashCard.java
@@ -10,7 +10,7 @@ public class FlashCard extends Question {
     String hint;
     public boolean trueFalse;
 
-    public FlashCard(String e, String i, String q, String a, int s, Date last, String sub, String top, String h) {
+    public FlashCard(String e, String i, String q, String a, double s, Date last, String sub, String top, String h) {
         super(e, i, q, a, s, last, sub, top);
         hint = h;
         trueFalse = false;
