+++ app/src/main/java/m1/nayak/m1/objects/MultipleChoice.java
@@ -10,7 +10,7 @@ public class MultipleChoice extends Question {
 
     public ArrayList<String> choices;
 
-    public MultipleChoice(String e, String i, String q, String a, ArrayList<String> c, int score, Date last, String sub, String top) {
+    public MultipleChoice(String e, String i, String q, String a, ArrayList<String> c, double score, Date last, String sub, String top) {
         super(e, i, q, a, score, last, sub, top);
         choices = c;
     }
